remove `mirror` helps in README
I believe I'm the only user of this function and I feel we don't need
this as it's much simpler to use `aws s3 sync` to do the same thing
in a more reliable way, hence here I want to retire this feature and
limited this to myself.

In another words, I don't think it is worth to advertise this feature.