Fix path for Windows when spaces in the file path
My computer has a space in the username. This means that when trying to run julia after installing with jill, I get the error
``` 
'C:\Users\Michael' is not recognized as an internal or external command,
operable program or batch file.
```

Hopefully by adding quotes around the path to julia, this fixes the issue.